Observe Activity lifecycle in MeshServiceClient to bind mesh service
