+++ app/src/main/java/com/geeksville/mesh/MainActivity.kt
@@ -44,14 +44,11 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.platform.LocalView
 import androidx.core.content.edit
 import androidx.core.net.toUri
-import com.geeksville.mesh.android.BindFailedException
 import com.geeksville.mesh.android.GeeksvilleApplication
 import com.geeksville.mesh.android.Logging
 import com.geeksville.mesh.model.BluetoothViewModel
 import com.geeksville.mesh.model.UIViewModel
 import com.geeksville.mesh.navigation.DEEP_LINK_BASE_URI
-import com.geeksville.mesh.service.MeshService
-import com.geeksville.mesh.service.startService
 import com.geeksville.mesh.ui.MainMenuAction
 import com.geeksville.mesh.ui.MainScreen
 import com.geeksville.mesh.ui.common.theme.AppTheme
@@ -70,6 +67,7 @@ class MainActivity :
     private val bluetoothViewModel: BluetoothViewModel by viewModels()
     private val model: UIViewModel by viewModels()
 
+    // This is aware of the Activity lifecycle and handles binding to the mesh service.
     @Inject internal lateinit var meshServiceClient: MeshServiceClient
 
     private var showAppIntro by mutableStateOf(false)
@@ -217,27 +215,6 @@ class MainActivity :
                 it.data?.data?.let { file_uri -> model.saveRangetestCSV(file_uri) }
             }
         }
-
-    private fun bindMeshService() {
-        debug("Binding to mesh service!")
-        try {
-            MeshService.startService(this)
-        } catch (ex: Exception) {
-            errormsg("Failed to start service from activity - but ignoring because bind will work ${ex.message}")
-        }
-
-        meshServiceClient.connect(this, MeshService.createIntent(), BIND_AUTO_CREATE + BIND_ABOVE_CLIENT)
-    }
-
-    override fun onStart() {
-        super.onStart()
-
-        try {
-            bindMeshService()
-        } catch (ex: BindFailedException) {
-            errormsg("Bind of MeshService failed${ex.message}")
-        }
-    }
 
     private fun showSettingsPage() {
         createSettingsIntent().send()
